Updating files for events and adding pictures
