--- src/data/eventData.ts
@@ -1,26 +1,0 @@
-export const events = [
-  {
-    name: "Event 1",
-    info: "Clubbing", //Fill with info about events
-    border: "border-csa-green-100",
-    text: "text-csa-green-100",
-  },
-  {
-    name: "Event 2",
-    info: "CHF",
-    border: "border-csa-yellow-100",
-    text: "text-csa-yellow-100",
-  },
-  {
-    name: "Event 3",
-    info: "Winter Retreat",
-    border: "border-csa-gray-100",
-    text: "text-csa-gray-100",
-  },
-  {
-    name: "Event 4",
-    info: "Banquet",
-    border: "border-csa-red-100",
-    text: "text-csa-red-100",
-  },
-];
